[Bug][183] Prevent tooltip from affecting page height.
The tooltip uses visibility to hide the element from view
so that we can support CSS transitions on visibility. The
combination of using visibility without having a position
set caused the tooltips to affect the height of the page.
By setting a position (top, left), we remove the tooltips
from the flow of the page.